requirements.txt
pytest==7.4.4
selenium==4.21.0
webdriver-manager==4.0.2
python-dotenv==1.0.1
requests==2.32.3
allure-pytest==2.13.5
undetected-chromedriver==3.5.5 
